org: spin out module and place utility in util file
