+++ index.html
@@ -1,14 +1,14 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="style.css" />
     <title>Dungeon Master's Friend</title>
-</head>
-<body>
+  </head>
+  <body>
     <h1>Dungeon Master's Friend</h1>
-    <script type='module'src="main.js"></script>
-</body>
+    <script type="module" src="main.js"></script>
+  </body>
 </html>
